Adds more documentation.
