+++ config/test-certs/generate_secrets.bash
@@ -19,6 +19,8 @@
 # limitations under the License.
 #
 
+# This script generates secrets with test certs for use in local testing.
+
 kubectl delete secrets -l app=fdb-kubernetes-operator
 kubectl create secret generic fdb-kubernetes-operator-secrets --from-file=config/test-certs/key.pem --from-file=config/test-certs/cert.pem
 kubectl patch secret fdb-kubernetes-operator-secrets --type='json' -p='[{"op": "add", "path": "/metadata/labels", "value":{"app":"fdb-kubernetes-operator"}}]'
