+++ pkg/controller/foundationdbcluster/pod_models.go
@@ -1,0 +1,290 @@
+/*
+ * pod_models.go
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2019 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package foundationdbcluster
+
+import (
+	"fmt"
+	"strconv"
+	"strings"
+
+	fdbtypes "github.com/foundationdb/fdb-kubernetes-operator/pkg/apis/apps/v1beta1"
+	corev1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/util/intstr"
+)
+
+// GetPodSpec builds a pod spec for a FoundationDB pod
+func GetPodSpec(cluster *fdbtypes.FoundationDBCluster, processClass string, podID string) *corev1.PodSpec {
+	imageName := cluster.Spec.MainContainer.ImageName
+	if imageName == "" {
+		imageName = "foundationdb/foundationdb"
+	}
+	podName := fmt.Sprintf("%s-%s", cluster.ObjectMeta.Name, podID)
+	mainContainer := corev1.Container{
+		Name:  "foundationdb",
+		Image: fmt.Sprintf("%s:%s", imageName, cluster.Spec.Version),
+		Env: []corev1.EnvVar{
+			corev1.EnvVar{Name: "FDB_CLUSTER_FILE", Value: "/var/dynamic-conf/fdb.cluster"},
+			corev1.EnvVar{Name: "FDB_TLS_CA_FILE", Value: "/var/dynamic-conf/ca.pem"},
+		},
+		Command: []string{"sh", "-c"},
+		Args: []string{
+			"fdbmonitor --conffile /var/dynamic-conf/fdbmonitor.conf" +
+				" --lockfile /var/fdb/fdbmonitor.lockfile",
+		},
+		Resources: *cluster.Spec.Resources,
+		VolumeMounts: []corev1.VolumeMount{
+			corev1.VolumeMount{Name: "data", MountPath: "/var/fdb/data"},
+			corev1.VolumeMount{Name: "dynamic-conf", MountPath: "/var/dynamic-conf"},
+			corev1.VolumeMount{Name: "fdb-trace-logs", MountPath: "/var/log/fdb-trace-logs"},
+		},
+	}
+
+	customizeContainer(&mainContainer, cluster.Spec.MainContainer)
+
+	sidecarEnv := make([]corev1.EnvVar, 0, 5)
+	sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "COPY_ONCE", Value: "1"})
+	sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "SIDECAR_CONF_DIR", Value: "/var/input-files"})
+	sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "FDB_PUBLIC_IP", ValueFrom: &corev1.EnvVarSource{
+		FieldRef: &corev1.ObjectFieldSelector{FieldPath: "status.podIP"},
+	}})
+
+	faultDomainKey := cluster.Spec.FaultDomain.Key
+	if faultDomainKey == "" {
+		faultDomainKey = "kubernetes.io/hostname"
+	}
+
+	faultDomainSource := cluster.Spec.FaultDomain.ValueFrom
+	if faultDomainSource == "" {
+		faultDomainSource = "spec.nodeName"
+	}
+
+	if faultDomainKey == "foundationdb.org/none" {
+		sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "FDB_MACHINE_ID", ValueFrom: &corev1.EnvVarSource{
+			FieldRef: &corev1.ObjectFieldSelector{FieldPath: "metadata.name"},
+		}})
+		sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "FDB_ZONE_ID", ValueFrom: &corev1.EnvVarSource{
+			FieldRef: &corev1.ObjectFieldSelector{FieldPath: "metadata.name"},
+		}})
+	} else if faultDomainKey == "foundationdb.org/kubernetes-cluster" {
+		sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "FDB_MACHINE_ID", ValueFrom: &corev1.EnvVarSource{
+			FieldRef: &corev1.ObjectFieldSelector{FieldPath: "spec.nodeName"},
+		}})
+		sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "FDB_ZONE_ID", Value: cluster.Spec.FaultDomain.Value})
+	} else {
+		sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "FDB_MACHINE_ID", ValueFrom: &corev1.EnvVarSource{
+			FieldRef: &corev1.ObjectFieldSelector{FieldPath: "spec.nodeName"},
+		}})
+		if !strings.HasPrefix(faultDomainSource, "$") {
+			sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "FDB_ZONE_ID", ValueFrom: &corev1.EnvVarSource{
+				FieldRef: &corev1.ObjectFieldSelector{FieldPath: faultDomainSource},
+			}})
+		}
+	}
+
+	instanceID := podID
+	if cluster.Spec.InstanceIDPrefix != "" {
+		instanceID = fmt.Sprintf("%s-%s", cluster.Spec.InstanceIDPrefix, instanceID)
+	}
+
+	sidecarEnv = append(sidecarEnv, corev1.EnvVar{Name: "FDB_INSTANCE_ID", Value: instanceID})
+
+	sidecarImageName := cluster.Spec.SidecarContainer.ImageName
+	if sidecarImageName == "" {
+		sidecarImageName = "foundationdb/foundationdb-kubernetes-sidecar"
+	}
+
+	initContainer := corev1.Container{
+		Name:  "foundationdb-kubernetes-init",
+		Image: fmt.Sprintf("%s:%s", sidecarImageName, cluster.GetFullSidecarVersion()),
+		Env:   sidecarEnv,
+		VolumeMounts: []corev1.VolumeMount{
+			corev1.VolumeMount{Name: "config-map", MountPath: "/var/input-files"},
+			corev1.VolumeMount{Name: "dynamic-conf", MountPath: "/var/output-files"},
+		},
+	}
+
+	customizeContainer(&initContainer, cluster.Spec.SidecarContainer)
+
+	sidecarEnv = append(sidecarEnv,
+		corev1.EnvVar{Name: "FDB_TLS_VERIFY_PEERS", Value: cluster.Spec.SidecarContainer.PeerVerificationRules})
+	sidecarEnv = append(sidecarEnv,
+		corev1.EnvVar{Name: "FDB_TLS_CA_FILE", Value: "/var/input-files/ca.pem"})
+
+	sidecarContainer := corev1.Container{
+		Name:  "foundationdb-kubernetes-sidecar",
+		Image: initContainer.Image,
+		Env:   sidecarEnv[1:],
+		VolumeMounts: []corev1.VolumeMount{
+			corev1.VolumeMount{Name: "config-map", MountPath: "/var/input-files"},
+			corev1.VolumeMount{Name: "dynamic-conf", MountPath: "/var/output-files"},
+		},
+		ReadinessProbe: &corev1.Probe{
+			Handler: corev1.Handler{
+				TCPSocket: &corev1.TCPSocketAction{
+					Port: intstr.IntOrString{IntVal: 8080},
+				},
+			},
+		},
+	}
+
+	if cluster.Spec.SidecarContainer.EnableTLS {
+		sidecarContainer.Args = []string{"--tls"}
+	}
+
+	customizeContainer(&sidecarContainer, cluster.Spec.SidecarContainer)
+
+	var mainVolumeSource corev1.VolumeSource
+	if usePvc(cluster, processClass) {
+		mainVolumeSource.PersistentVolumeClaim = &corev1.PersistentVolumeClaimVolumeSource{
+			ClaimName: fmt.Sprintf("%s-data", podName),
+		}
+	} else {
+		mainVolumeSource.EmptyDir = &corev1.EmptyDirVolumeSource{}
+	}
+
+	volumes := []corev1.Volume{
+		corev1.Volume{Name: "data", VolumeSource: mainVolumeSource},
+		corev1.Volume{Name: "dynamic-conf", VolumeSource: corev1.VolumeSource{EmptyDir: &corev1.EmptyDirVolumeSource{}}},
+		corev1.Volume{Name: "config-map", VolumeSource: corev1.VolumeSource{ConfigMap: &corev1.ConfigMapVolumeSource{
+			LocalObjectReference: corev1.LocalObjectReference{Name: fmt.Sprintf("%s-config", cluster.Name)},
+			Items: []corev1.KeyToPath{
+				corev1.KeyToPath{Key: fmt.Sprintf("fdbmonitor-conf-%s", processClass), Path: "fdbmonitor.conf"},
+				corev1.KeyToPath{Key: "cluster-file", Path: "fdb.cluster"},
+				corev1.KeyToPath{Key: "ca-file", Path: "ca.pem"},
+				corev1.KeyToPath{Key: "sidecar-conf", Path: "config.json"},
+			},
+		}}},
+		corev1.Volume{Name: "fdb-trace-logs", VolumeSource: corev1.VolumeSource{EmptyDir: &corev1.EmptyDirVolumeSource{}}},
+	}
+
+	for _, volume := range cluster.Spec.Volumes {
+		volumes = append(volumes, *volume.DeepCopy())
+	}
+
+	var affinity *corev1.Affinity
+
+	if faultDomainKey != "foundationdb.org/none" && faultDomainKey != "foundationdb.org/kubernetes-cluster" {
+		affinity = &corev1.Affinity{
+			PodAntiAffinity: &corev1.PodAntiAffinity{
+				PreferredDuringSchedulingIgnoredDuringExecution: []corev1.WeightedPodAffinityTerm{
+					corev1.WeightedPodAffinityTerm{
+						Weight: 1,
+						PodAffinityTerm: corev1.PodAffinityTerm{
+							TopologyKey: faultDomainKey,
+							LabelSelector: &metav1.LabelSelector{MatchLabels: map[string]string{
+								"fdb-cluster-name":  cluster.ObjectMeta.Name,
+								"fdb-process-class": processClass,
+							}},
+						},
+					},
+				},
+			},
+		}
+	}
+
+	initContainers := []corev1.Container{initContainer}
+	initContainers = append(initContainers, cluster.Spec.InitContainers...)
+
+	containers := []corev1.Container{mainContainer, sidecarContainer}
+	containers = append(containers, cluster.Spec.Containers...)
+
+	return &corev1.PodSpec{
+		InitContainers:               initContainers,
+		Containers:                   containers,
+		Volumes:                      volumes,
+		Affinity:                     affinity,
+		SecurityContext:              cluster.Spec.PodSecurityContext,
+		AutomountServiceAccountToken: cluster.Spec.AutomountServiceAccountToken,
+	}
+}
+
+// usePvc determines whether we should attach a PVC to a pod.
+func usePvc(cluster *fdbtypes.FoundationDBCluster, processClass string) bool {
+	return cluster.Spec.VolumeSize != "0" && isStateful(processClass)
+}
+
+// isStateful determines whether a process class should store data.
+func isStateful(processClass string) bool {
+	return processClass == "storage" || processClass == "log" || processClass == "transaction"
+}
+
+// GetPvc builds a persistent volume claim for a FoundationDB instance.
+func GetPvc(cluster *fdbtypes.FoundationDBCluster, processClass string, id int) (*corev1.PersistentVolumeClaim, error) {
+	if !usePvc(cluster, processClass) {
+		return nil, nil
+	}
+	name := fmt.Sprintf("%s-%d-data", cluster.ObjectMeta.Name, id)
+	size, err := resource.ParseQuantity(cluster.Spec.VolumeSize)
+	if err != nil {
+		return nil, err
+	}
+	spec := corev1.PersistentVolumeClaimSpec{
+		AccessModes: []corev1.PersistentVolumeAccessMode{corev1.ReadWriteOnce},
+		Resources: corev1.ResourceRequirements{
+			Requests: corev1.ResourceList{"storage": size},
+		},
+		StorageClassName: cluster.Spec.StorageClass,
+	}
+
+	var idLabel string
+	if id > 0 {
+		idLabel = strconv.Itoa(id)
+	} else {
+		idLabel = ""
+	}
+
+	return &corev1.PersistentVolumeClaim{
+		ObjectMeta: metav1.ObjectMeta{
+			Namespace: cluster.Namespace,
+			Name:      name,
+			Labels:    getPodLabels(cluster, processClass, idLabel),
+		},
+		Spec: spec,
+	}, nil
+}
+
+// customizeContainer adds container overrides from the cluster spec to a
+// container.
+func customizeContainer(container *corev1.Container, overrides fdbtypes.ContainerOverrides) {
+	envOverrides := make(map[string]bool)
+	fullEnv := []corev1.EnvVar{}
+
+	for _, envVar := range overrides.Env {
+		fullEnv = append(fullEnv, *envVar.DeepCopy())
+		envOverrides[envVar.Name] = true
+	}
+
+	for _, envVar := range container.Env {
+		if !envOverrides[envVar.Name] {
+			fullEnv = append(fullEnv, envVar)
+		}
+	}
+
+	container.Env = fullEnv
+
+	for _, volume := range overrides.VolumeMounts {
+		container.VolumeMounts = append(container.VolumeMounts, *volume.DeepCopy())
+	}
+
+	container.SecurityContext = overrides.SecurityContext
+}
