+++ pkg/controller/foundationdbcluster/set_default_values.go
@@ -1,3 +1,23 @@
+/*
+ * set_default_values.go
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2019 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package foundationdbcluster
 
 import (
@@ -9,6 +29,8 @@ import (
 	"k8s.io/apimachinery/pkg/api/resource"
 )
 
+// SetDefaultValues provides a reconciliation step for setting default values in
+// the cluster spec.
 type SetDefaultValues struct {
 }
