+++ pkg/controller/foundationdbcluster/update_config_map.go
@@ -1,3 +1,23 @@
+/*
+ * update_config_map.go
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2019 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package foundationdbcluster
 
 import (
@@ -11,6 +31,8 @@ import (
 	"k8s.io/apimachinery/pkg/types"
 )
 
+// UpdateConfigMap provides a reconciliation step for updating the dynamic conf
+// for a cluster.
 type UpdateConfigMap struct{}
 
 func (u UpdateConfigMap) Reconcile(r *ReconcileFoundationDBCluster, context ctx.Context, cluster *fdbtypes.FoundationDBCluster) (bool, error) {
