+++ pkg/controller/foundationdbcluster/update_sidecar_versions.go
@@ -1,3 +1,23 @@
+/*
+ * update_sidecar_versions.go
+ *
+ * This source file is part of the FoundationDB open source project
+ *
+ * Copyright 2019 Apple Inc. and the FoundationDB project authors
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
 package foundationdbcluster
 
 import (
@@ -8,6 +28,8 @@ import (
 	fdbtypes "github.com/foundationdb/fdb-kubernetes-operator/pkg/apis/apps/v1beta1"
 )
 
+// UpdateSidecarVersions provides a reconciliation step for upgrading the
+// sidecar.
 type UpdateSidecarVersions struct {
 }
